+++ server/src/metrics_collector_controllers/collector_utils.rs
@@ -28,7 +28,7 @@ pub struct Proc {
     pub proc_name: String,
     pub num_threads: i64,
     pub proc_mem: String,
-    pub proc_cpu: u64
+    pub proc_cpu: f32
 }
 
 impl Default for Proc {
@@ -39,14 +39,14 @@ impl Default for Proc {
             proc_name: "".to_owned(),
             num_threads: 0,
             proc_mem: "".to_owned(),
-            proc_cpu: 0
+            proc_cpu: 0.0
         }
     }
 }
 
 impl Proc {
     // Construct process
-    pub fn new(uuid: String, id: i32, name: &str, threads: i64, mem: &str, cpu: u64) -> Proc {
+    pub fn new(uuid: String, id: i32, name: &str, threads: i64, mem: &str, cpu: f32) -> Proc {
         Proc {
             uuid: Uuid::new_v4().to_string(),
             proc_id: id,
@@ -73,7 +73,7 @@ impl Proc {
         self.proc_mem = memory.to_string();
     }
 
-    pub fn set_cpu_usage(&mut self, cpu_usage: u64) { self.proc_cpu = cpu_usage }
+    pub fn set_cpu_usage(&mut self, cpu_usage: f32) { self.proc_cpu = cpu_usage }
 }
 
 /*
